Fix AI error description export
